I should not be trusted with merge privileges
Well there you go, I got Kiri's stuff in but while I was trying to do it I closed the project to not get any weird results and game maker interpreted that as clearing all of the asset imports on it's side but not on githubs side, ergo none of the assets were visible to game maker but they still existed on github. Fun! Now I thought that I could just undo the changes on github and do it again, but github cries anytime you try to undo things so I ended up just "fixing" the project with yyp fixer or whatever and pasting it back in as a replacement. Basically, I deleted and recopied the entire project hence why there's so many things. Whoop de doo.