Fixed for real this time?
There was an ordering issue that made doing anything with orbs crash, that's fun. But I also fixed a few other things and cleared a boss template so that making new bosses in the future is just a matter of giving them ai and specific states and everything else is auto done / inherited. This should make it very simple to make new bosses once we know what the moves and visuals are supposed to be.