Spider boss complete and odds and ends
The spider boss now has interesting ai and multiple attacks and states, both bosses have death states and the loops for them work. The boss selector in the menu now works and many odds and ends have been fixed / updated to make the bosses more proper and work with everything.