Added transition area, not done, plus others
There is now an area that opens up once the boss is dead and you can walk down and take a boat to the next level. Lots of things still to add but this is important. Also changed some stuff around and added placeholder sprites that I hope can be made for the rollie pollie boss