Expanded level loop, orb types, overall polish
Things have been changed, the levels are now cyclical where riding the raft takes you to the next level and each level has an entry area, a boss fight, and an exit on the raft. I also made the 3 orb types, webs that slow down but do nothing else, ice that slows down for a time and does some damage when you form a ring (ice and fire are not moved by the boss, very convenient) and fire that just does a bunch of damage in the ring. I also fixed up / completed most of the game loop in terms of ai and spawning and some random other things. Mantis still not in and roller still silly looking..