Pause menu now done, spawn outside of fight now
The main thing here (besides trying to get fmod to work for like 3 hours, unsuccessfully) is the pause menu, just hit escape to open it and observe that. I also adjusted the spawn so now when you join the game it put's you on the left shore and you have to enter the boss area to start the fight, then when you win you sail to the next level and rinse and repeat.