Bunch of smallish changes to game play + intros
The roller boss and spider now have basic intros and I reworked a bunch of things to make attacks and ai nicer, the roller still needs work but the spider at least functions and is interesting. Animation has been changed for spider and lot's of states have been added to allow for more distinct behaviors in roller especially. I tried my best to give the attacks some animations but they're very crude for now. New attacks were added.
The player can also now dodge into orbs to push them around and I fixed some bugs related to state setting. The final music tracks are now in and roughly play correctly, they still need to be faded out while exiting bosses but I'm waiting for the river track. You will also note that all the map assets are gone which is the result of some issues with scaling and we're waiting to redraw the map properly and hopefully get that back in looking correct. Sounds are also pretty lacking besides music.. I hate trying to find sounds and we haven't made many ourselves. That's the state at the moment, needs more animation work and adjustments. The mantis still isn't in at all but it should be relatively easy to throw together in the next couple days. Regardless, later